Use phantomjs instead of chromedriver
This means we don't have to mess with something like SauceLabs
in Travis CI.

I had to work around a strange bug, though. It seems that in the latest
version of selenium-webdriver, the phantomjs harness is broken unless
you explicitly set webdriver.logging.LevelName to something. This is
silly and I hope we can fix it sometime soon.